+++ core-mods/debugaddopt.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python3
 # Call it in the platform folder without an argument,
-# then boards.txt will be renamed to boards.txt.backup and boards.txt gets modified
-# If it is called with an argument, then it is expected to be the boards.txt file,
-# the file is renamed to *.backup and the original file is modified
+# then boards.txt will be renamed to boards.txt.backup and boards.txt gets modified.
 # The modification adds a menu entry for each MCU/board called "Debug Compile Flags" with
 # four possible values "No Debug", "Debug", "Debug (no LTO)", and "Debug (no LTO, no optimizations)", which will add
 # flags to .build.extra_flags
-
+# Similarly, platform.txt is modfied. A few lines are added to allow debugging under Arduino IDE 2.X 
 
 import os, sys
+
+exception = ['8', 'attiny26'] # these boards do not supprt debugWIRE
 
 def massage(lines):
     allboards = gatherboards(lines)
@@ -21,15 +21,18 @@ def massage(lines):
             newlines.append("\n");
             newlines.append(b + ".menu.debug.os=No Debug\n")
             newlines.append(b + ".menu.debug.os.build.debug=\n")
-            newlines.append(b + ".menu.debug.og=Debug\n")
-            newlines.append(b + ".menu.debug.og.build.debug=-Og\n")
-            newlines.append(b + ".menu.debug.o0=Debug (no comp. optim.)\n")
-            newlines.append(b + ".menu.debug.o0.build.debug=-O0\n")
-            newlines.append(b + ".menu.debug.no-lto=Debug (no LTO)\n")
-            newlines.append(b + ".menu.debug.no-lto.build.debug=-Og -fno-lto\n")
-            newlines.append(b + ".menu.debug.no-ltoo0=Debug (no LTO, no comp. optim.)\n")
-            newlines.append(b + ".menu.debug.no-ltoo0.build.debug=-O0 -fno-lto\n")
-            newlines.append(b + ".build.extra_flags={build.debug} " + extraflags.get(b,"\n"))
+            if b in exception:
+                newlines.append(b + ".debug.executable=\n")
+            else:
+                newlines.append(b + ".menu.debug.og=Debug\n")
+                newlines.append(b + ".menu.debug.og.build.debug=-g3 -Og\n")
+                newlines.append(b + ".menu.debug.o0=Debug (no comp. optim.)\n")
+                newlines.append(b + ".menu.debug.o0.build.debug=-g3 -O0\n")
+                newlines.append(b + ".menu.debug.no-lto=Debug (no LTO)\n")
+                newlines.append(b + ".menu.debug.no-lto.build.debug=-g3 -Og -fno-lto\n")
+                newlines.append(b + ".menu.debug.no-ltoo0=Debug (no LTO, no comp. optim.)\n")
+                newlines.append(b + ".menu.debug.no-ltoo0.build.debug=-g3 -O0 -fno-lto\n")
+                newlines.append(b + ".build.extra_flags={build.debug} " + extraflags.get(b,"\n"))
     return newlines
 
 def gatherboards(lines):
@@ -49,28 +52,86 @@ def gatherextra(lines, boards):
     return d
 
 
-if len(sys.argv) > 2:
-    print("Call script with one or none argument only")
-    exit()
-if len(sys.argv) == 2:
-    filnam = sys.argv[1]
-else:
-    filnam = "boards.txt"
+filnam = "boards.txt"
 
 if not os.path.isfile(filnam):
     print(filnam, "does not exist")
     exit()
-os.rename(filnam, filnam + ".backup")
-file = open(filnam + ".backup")
+file = open(filnam)
 alllines = file.readlines()
 file.close()
 
+if "#################### DEBUG FLAG HANDLING ####################\n" in alllines:
+    print("Board file already contains debug flag handling")
+    exit(1)
+
+os.rename(filnam, filnam + ".backup")
 newlines = massage(alllines)
 
 file = open(filnam, "w")
 file.writelines(alllines + newlines)
 file.close()
 if len(newlines) == 0:
-    print("No boards found, no lines added")
+    print("*** No boards found, no lines added")
 else:
     print("Boards file successfully modified")
+
+filnam = "platform.txt"
+
+if not os.path.isfile(filnam):
+    print(filnam, "does not exist")
+    exit()
+file = open(filnam)
+alllines = file.readlines()
+file.close()
+
+if "# EXPERIMENTAL feature: optimization flags\n" in alllines:
+    print("Platform file already contains debug optimization flags\n")
+    exit(2)
+
+os.rename(filnam, filnam + ".backup")
+
+lastline = ""
+optinserted = False
+file = open(filnam, "w")
+for line in alllines:
+    if optinserted and "-Os" in line:
+        line = line[:line.find("-Os")] + " {compiler.optimization_flags} " + line[line.find("-Os")+3:]
+    file.write(line)
+    if "# ---------------" in line and "# AVR compile variables" in lastline:
+        file.write("\n")
+        file.write("# EXPERIMENTAL feature: optimization flags\n")
+        file.write("#  - this is alpha and may be subject to change without notice\n")
+        file.write("compiler.optimization_flags=-Os\n")
+        file.write("compiler.optimization_flags.release=-Os\n")
+        file.write("compiler.optimization_flags.debug=-Og -g3\n")
+        optinserted = True
+    lastline = line
+
+file.write("\n")
+
+if optinserted:
+    print("Found place to insert debug compiler optimization flags")
+else:
+    print("*** Did not find place to insert debug compiler optimization flags")
+    exit(1)
+
+file.write("# Debugger configuration (general options)\n")
+file.write("# ----------------------------------------\n")
+file.write("# EXPERIMENTAL feature:\n")
+file.write("#  - this is alpha and may be subject to change without notice\n")
+file.write("debug.executable={build.path}/{build.project_name}.elf\n")
+file.write("debug.toolchain=gcc\n")
+file.write("debug.toolchain.path={runtime.tools.dw-link-tools.path}\n")
+file.write("\n")
+file.write("debug.server=openocd\n")
+file.write("debug.server.openocd.path={debug.toolchain.path}/dw-server\n")
+file.write("#doesn't matter, but should be specified so that cortex-debug is happy\n")
+file.write("debug.server.openocd.scripts_dir={debug.toolchain.path}/\n")
+file.write("#doesn't matter, but should be specified so that cortex-debug is happy\n")
+file.write("debug.server.openocd.script={debug.toolchain.path}/dw-server\n")
+
+print("Platform file successfully modified")
+
+file.close()
+